+++ webrtc/modules/audio_coding/neteq/dtmf_buffer.cc
@@ -55,7 +55,7 @@ namespace webrtc {
 //
 int DtmfBuffer::ParseEvent(uint32_t rtp_timestamp,
                            const uint8_t* payload,
-                           size_t payload_length_bytes,
+                           int payload_length_bytes,
                            DtmfEvent* event) {
   if (!payload || !event) {
     return kInvalidPointer;
