+++ webrtc/modules/rtp_rtcp/interface/fec_receiver.h
@@ -24,7 +24,7 @@ class FecReceiver {
 
   virtual int32_t AddReceivedRedPacket(const RTPHeader& rtp_header,
                                        const uint8_t* incoming_rtp_packet,
-                                       size_t packet_length,
+                                       int packet_length,
                                        uint8_t ulpfec_payload_type) = 0;
 
   virtual int32_t ProcessReceivedFec() = 0;
