+++ webrtc/modules/rtp_rtcp/interface/rtp_receiver.h
@@ -72,7 +72,7 @@ class RtpReceiver {
   // detected and acted upon.
   virtual bool IncomingRtpPacket(const RTPHeader& rtp_header,
                                  const uint8_t* payload,
-                                 size_t payload_length,
+                                 int payload_length,
                                  PayloadUnion payload_specific,
                                  bool in_order) = 0;
