+++ webrtc/modules/rtp_rtcp/source/nack_rtx_unittest.cc
@@ -40,7 +40,7 @@ class VerifyingRtxReceiver : public NullRtpData
 
   virtual int32_t OnReceivedPayloadData(
       const uint8_t* data,
-      const size_t size,
+      const uint16_t size,
       const webrtc::WebRtcRTPHeader* rtp_header) OVERRIDE {
     if (!sequence_numbers_.empty())
       EXPECT_EQ(kTestSsrc, rtp_header->header.ssrc);
@@ -95,7 +95,7 @@ class RtxLoopBackTransport : public webrtc::Transport {
     packet_loss_ = 0;
   }
 
-  virtual int SendPacket(int channel, const void *data, size_t len) OVERRIDE {
+  virtual int SendPacket(int channel, const void *data, int len) OVERRIDE {
     count_++;
     const unsigned char* ptr = static_cast<const unsigned  char*>(data);
     uint32_t ssrc = (ptr[8] << 24) + (ptr[9] << 16) + (ptr[10] << 8) + ptr[11];
@@ -105,13 +105,13 @@ class RtxLoopBackTransport : public webrtc::Transport {
         sequence_number);
     if (packet_loss_ > 0) {
       if ((count_ % packet_loss_) == 0) {
-        return static_cast<int>(len);
+        return len;
       }
     } else if (count_ >= consecutive_drop_start_ &&
         count_ < consecutive_drop_end_) {
-      return static_cast<int>(len);
+      return len;
     }
-    size_t packet_length = len;
+    int packet_length = len;
     // TODO(pbos): Figure out why this needs to be initialized. Likely this
     // is hiding a bug either in test setup or other code.
     // https://code.google.com/p/webrtc/issues/detail?id=3183
@@ -143,14 +143,12 @@ class RtxLoopBackTransport : public webrtc::Transport {
                                           true)) {
       return -1;
     }
-    return static_cast<int>(len);
+    return len;
   }
 
-  virtual int SendRTCPPacket(int channel,
-                             const void *data,
-                             size_t len) OVERRIDE {
+  virtual int SendRTCPPacket(int channel, const void *data, int len) OVERRIDE {
     if (module_->IncomingRtcpPacket((const uint8_t*)data, len) == 0) {
-      return static_cast<int>(len);
+      return len;
     }
     return -1;
   }
@@ -216,7 +214,7 @@ class RtpRtcpRtxNackTest : public ::testing::Test {
                                                        0,
                                                        video_codec.maxBitrate));
 
-    for (size_t n = 0; n < payload_data_length; n++) {
+    for (int n = 0; n < payload_data_length; n++) {
       payload_data[n] = n % 10;
     }
   }
@@ -294,7 +292,7 @@ class RtpRtcpRtxNackTest : public ::testing::Test {
   RtxLoopBackTransport transport_;
   VerifyingRtxReceiver receiver_;
   uint8_t  payload_data[65000];
-  size_t payload_data_length;
+  int payload_data_length;
   SimulatedClock fake_clock;
 };
